# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GoalGetter API",
    "description": "API for the GoalGetter app",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/goals/objective-questions": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Objective Questions",
        "description": "Step 1: validate the prompt is a real goal, then generate clarifying\nmultiple-choice questions. Each question names the model that wrote it, and\nthe app sends that back with its answer to `POST /goals` (#216).\n\nPublic, so there is no student row: the language is the header the app\nsends (#172), else the one the prompt is written in (#173).",
        "operationId": "objective_questions_api_v1_goals_objective_questions_post",
        "parameters": [
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectiveQuestionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObjectiveQuestion"
                  },
                  "title": "Response Objective Questions Api V1 Goals Objective Questions Post"
                }
              }
            }
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/study-plan": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Study Plan",
        "description": "Step 2: generate a stateless study-plan preview (goal name + markdown\ndescription) from the prompt and the user's onboarding answers. Not persisted.",
        "operationId": "study_plan_api_v1_goals_study_plan_post",
        "parameters": [
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyPlanResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Create Goal",
        "description": "Step 3 (AUTHED): persist the goal the user approved in the preview, store the\nonboarding it came from, make it the active goal, then fire the student chain\nin the background. No Gemini call of its own (#132): the wait it used to buy\nintroduction screens for is now the standard questions this returns.\n\nThe onboarding is *stored*, not handed to the chain: the chain's first step\nreads it from the database, so a run that fails is one the nightly run can do\nover (#88). It is read **as it stood when this returned**, which is what keeps\nthe batch now in flight from seeing the standard questions the student is\nabout to answer.",
        "operationId": "create_goal_api_v1_goals_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCommitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalCreationResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "List Goals",
        "description": "Every goal of the student, newest first, with all the fields the goals list and\nthe goal detail screen show (there is no per-goal GET).",
        "operationId": "list_goals_api_v1_goals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalResponse"
                  },
                  "title": "Response List Goals Api V1 Goals Get"
                }
              }
            }
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}/standard-answers": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Standard Answers",
        "description": "What the student told us about himself while his first batch generated.\n\nStored beside the rest of his onboarding, marked `ai_model = \"system\"`. The\nbatch already in flight will not read them - `POST /goals` pinned its view of\nthe onboarding to the moment it returned - and every generation after it\nwill (#132). Answers so far are worth keeping, so a partial list is normal\nand an empty one is a no-op.",
        "operationId": "standard_answers_api_v1_goals__goal_id__standard_answers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardAnswersRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}/set-active": {
      "put": {
        "tags": [
          "goals"
        ],
        "summary": "Set Active Goal",
        "description": "Make one of the student's goals the active one (students.current_goal_id).\nSomeone else's goal, or a missing one, is 404 (see get_owned_goal).",
        "operationId": "set_active_goal_api_v1_goals__goal_id__set_active_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetActiveGoalResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}": {
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Delete Goal",
        "description": "Delete a goal and everything under it. The database does the cascading\n(ON DELETE CASCADE on the goal's rows, SET NULL on students.current_goal_id), so\nthe in-memory student is refreshed afterwards: it may still hold the old id.",
        "operationId": "delete_goal_api_v1_goals__goal_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "x-student-language",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The student's chosen language: en, pt, es, fr or de",
              "title": "X-Student-Language"
            },
            "description": "The student's chosen language: en, pt, es, fr or de"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "description": "The request was refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "The server or a service it calls failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "invalid_request",
          "route_not_found",
          "method_not_allowed",
          "too_many_requests",
          "internal_error",
          "not_signed_in",
          "invalid_token",
          "student_no_longer_exists",
          "invalid_refresh_token",
          "invalid_google_token",
          "student_not_found",
          "no_active_goal",
          "goal_not_found",
          "not_a_goal",
          "lessons_not_ready",
          "unknown_question",
          "message_not_found",
          "google_unreachable",
          "gemini_no_answer",
          "gemini_key_rejected",
          "gemini_quota_exhausted",
          "gemini_failed",
          "gemini_timed_out",
          "gemini_unreachable"
        ],
        "title": "ErrorCode",
        "description": "The `code` of an error response. See the module docstring."
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong. The app switches on this."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "The same, in English, for a person reading a log."
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error the API answers (#214)."
      },
      "GoalCommitRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Prompt",
            "description": "What the user wants to learn"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveAnswer"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Answers",
            "description": "Answers to the objective questions"
          },
          "goal_name": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Goal Name",
            "description": "The approved goal name from the preview"
          },
          "description": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Description",
            "description": "The approved markdown description from the preview"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "answers",
          "goal_name",
          "description"
        ],
        "title": "GoalCommitRequest",
        "description": "Step 3 (POST /goals): commit the goal the user approved in the preview. Carries\nthe approved goal_name + description so we persist exactly what they saw (no\nre-generation), plus the prompt/answers the async jobs need for context."
      },
      "GoalCreationRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Prompt",
            "description": "What the user wants to learn"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveAnswer"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Answers",
            "description": "Answers to the objective questions"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "answers"
        ],
        "title": "GoalCreationRequest",
        "description": "Shared by step 2 (study-plan preview) and step 3 (create goal): the prompt\nplus the user's answers to the objective questions."
      },
      "GoalCreationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The created goal's id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The goal name"
          },
          "standard_questions": {
            "items": {
              "$ref": "#/components/schemas/StandardQuestionData"
            },
            "type": "array",
            "title": "Standard Questions",
            "description": "Questions to ask while the first batch generates"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "standard_questions"
        ],
        "title": "GoalCreationResponse",
        "description": "What the client gets after committing a goal: the persisted goal plus the\nstandard questions to ask while the chain generates its first batch (#132)."
      },
      "GoalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "current_elo": {
            "type": "integer",
            "title": "Current Elo",
            "description": "The student's rating for this goal (goals.rating)"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether this is the student's active goal (students.current_goal_id)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Bumped whenever the goal row changes, e.g. the rating after a lesson"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "current_elo",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "GoalResponse",
        "description": "One goal in GET /goals. Carries every field the goals list and the goal detail\nscreen show, so the detail screen needs no fetch of its own."
      },
...
      "ObjectiveQuestion": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options",
            "description": "Exactly 4 options"
          },
          "ai_model": {
            "type": "string",
            "title": "Ai Model",
            "description": "The model that wrote this question; sent back on its answer"
          }
        },
        "type": "object",
        "required": [
          "question",
          "options",
          "ai_model"
        ],
        "title": "ObjectiveQuestion",
        "description": "A clarifying multiple-choice question with exactly 4 options (no 'correct' one — these profile the user)."
      },
      "ObjectiveQuestionsRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Prompt",
            "description": "What the user wants to learn"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "ObjectiveQuestionsRequest",
        "description": "Step 1 of goal creation: the raw thing the user says they want to learn."
      },
...
      "SetActiveGoalResponse": {
        "properties": {
          "goal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Goal Id"
          }
        },
        "type": "object",
        "required": [
          "goal_id"
        ],
        "title": "SetActiveGoalResponse",
        "description": "PUT /goals/{goal_id}/set-active: the goal that is now active."
      },
...
      "StandardAnswersRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/StandardAnswer"
            },
            "type": "array",
            "title": "Answers",
            "description": "The answers given, in any order"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "StandardAnswersRequest",
        "description": "What the student answered before his first lesson. Partial by design: he\nmay skip out at any question, and what he did answer is still worth keeping."
      },
...
      "StudyPlanResponse": {
        "properties": {
          "goal_name": {
            "type": "string",
            "title": "Goal Name",
            "description": "Short, clear name for what the user will learn"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Markdown: the next thing to study and why"
          }
        },
        "type": "object",
        "required": [
          "goal_name",
          "description"
        ],
        "title": "StudyPlanResponse",
        "description": "Stateless preview the user reviews before committing to the goal."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}